+++ NEWS.md
@@ -2,12 +2,18 @@ vegtable2 0.1.2
 ===============
 
 ### New Features
+
 * New function `report_communities()`
 * New function `import_sudamerica()`
 * New function `version_svg()`
 * New function `get_description()`
 
+### Improvements
+
+* In function `postgres2vegtable()` import of header using `pgGetGeom()` from package `rpostgis`.
+
 ### Bug Fixes
+
 * Function `get_precision()` provided wrong information for short strings with NAs
 
 vegtable2 0.1.1